Atualização do conector MySQL para a versão 8.0.33

+++ src/ProdutosDTO.java
@@ -10,7 +10,7 @@
 public class ProdutosDTO {
     private Integer id;
     private String nome;
-    private Integer valor;
+    private double valor;
     private String status;
 
     public Integer getId() {
@@ -21,7 +21,7 @@ public class ProdutosDTO {
         return nome;
     }
 
-    public Integer getValor() {
+    public double getValor() {
         return valor;
     }
 
@@ -37,7 +37,7 @@ public class ProdutosDTO {
         this.nome = nome;
     }
 
-    public void setValor(Integer valor) {
+    public void setValor(double valor) {
         this.valor = valor;
     }
